Create files for ManOpenURLHandlerCommand class.
Add header and source files for the ManOpenURLHandlerCommand class,
moving it out of ManDocumentController.m.

Remove the `static' modifier from the IsSectionWord() function in
ManDocumentController.m and add an extern prototype for it to
ManDocumentController.h so that it can be called by both
ManDocumentController and ManOpenURLHandlerCommand.

Move the large comment block from the -performDefaultImplementation
method of ManOpenURLHandlerCommand and put it before the
ManOpenURLHandlerCommand @interface.