Add unit tests for ManOpenURLHandlerCommand.
Created the ManOpenTests unit test target.

Added initial test suite ManOpenURLHandlerCommandTests, with tests to
cover the -performDefaultImplementation method of
ManOpenURLHandlerCommand for the `x-man-page' URL scheme.

In the ManOpenURLHandlerCommand class, replaced the use of
[ManDocumentController sharedDocumentController] with the equivalent
manDocumentController property, which is initialized to
sharedDocumentController by default, but can be overridden in tests.

Created MockManDocumentController for testing to verify the -openString:
call in -performDefaultImplementation.