Handle `apropos' version of `x-man-page:' URLs.
Extend the ManOpenURLHandlerCommand to handle the `apropos' version of
`x-man-page:' URLs as handled by Terminal.app.  If the URL ends in
`:type=a', open an AproposDocument instead of a ManDocument.

Add unit tests to cover the new URL form, verifying that
ManOpenURLHandlerCommand calls -openApropos: instead of -openString: on
ManDocumentController.

Add GetURLForApropos.scpt script to demonstrate apropos URLs.