Add ManPage class to hold a section and page name.
The ManPage class is a container for an optional section and man page
name.  It included the class method +isSection:, which replaces the
function IsSectionWord() that was defined in ManDocumentController.m.

Refactored the handling of x-man-page: scheme URLs in the
-performDefaultImplementation method of ManOpenURLHandlerCommand.

Updated ManOpenURLHandlerCommandTests to account for a change in how
open strings are formatted. Added test cases for man page names and
apropos keywords that begin with a digit, and adjusted
ManOpenURLHandlerCommand behavior so that such names and keywords are
not mistaken for sections.