Enable manopen: URLs in ManOpenURLHandlerCommand.
In the -performDefaultImplementation method, check for and handle
manopen: URLs.  Also add missing -autorelease calls and reorder calls in
the -dealloc method.

Add the manopen: URL scheme to the Info.plist URL types, and add missing
Viewer role for the x-man-page: scheme.

Update the description for the GetURL command in
ManOpen.scriptTerminology to include manopen: and file: schemes.

Add example Applescripts for the manopen: scheme, and rename the
existing x-man-page: scheme example scripts.