Created Version class for comparing bundle versions.
The Version class is able to parse version strings in the form "1.2.3"
so that app bundle versions can be compared for ordering.

The Version class is part of the openman command line tool target.

Added the openmanTests unit test target.  Changed the ManOpen and
openman schemes to run the openmanTests target during the test action.