Fix deprecation warnings for old NSView dragging methods.
In the PoofDragTableView class, remove deprecated
-dragImage:at:offset:event:pasteboard:source:slideBack: method, since it
only contained a workaround for OS X Lion 10.7, which we no longer
support.

Also replace the -draggedImage:endedAt:operation: method with the
NSDraggingSource protocol method
-draggingSession:endedAtPoint:operation: and simplify the logic by
returning early for each unmet condition.

Update the -draggingSession:movedToPoint: method to use the
`animatesToStartingPositionsOnCancelOrFail' property on the
NSDraggingSession argument directly instead of through -setValue:forKey:
since this property is available on all versions of macOS that we run
on, and remove the `SessionSlideBackKey' #define that held the property
name.