Suppress "warning: code will never be executed" in generated code.
Both _cat2rtf_ and _cat2html_ executables are generated from `lex` files
with the `%option noyywrap` set, which causes the `yywrap()` function to
be #defined as 1.  The generated lexer contains `if (yywrap()) ... else
...` which trips the `-Wunreachable-code` warning.

Suppress this warning for _cat2rtf_ and _cat2html_ by adding
`-Wno-unreachable-code` to the "Other Warning Flags" for these targets.